requirements.txt
tensorflow-cpu
fastapi
Pillow
starlette
gunicorn
uvicorn
numpy
python-multipart
typing-extensions
opencv-python 
